Provide ig::server functions
This provides Puppet functions as wrappers around the existing Ruby
functions we have.  The goal of these is to be able to use Puppet
features like namespaces and datatypes.  We mostly use those functions
on the project modules.  Namespaces would make it obvious where they
are coming from.  Datatypes checks will save us from mistakes.

The new functions are also designed to be more convenient.  The current
multiadd function check if the attribute value is already there, so
the caller doesn't need to do it.