Provide function to update single attributes
This function allows to update single attributes from Puppet and is the
addition to its counter part adminapi::change_multi_attribute for multi
attributes.

Unfortunately it requires and additional request to determine the
current value which is necessary for making a valid commit to apply the
changes to Serveradmin but this should be fine if we use it
economically.